+++ lending/src/sdk/companyinfo.ts
@@ -107,7 +107,7 @@ export class CompanyInfo {
                     );
                 }
                 break;
-            case [401, 404, 409, 429].includes(httpRes?.status):
+            case [401, 402, 403, 404, 409, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -131,9 +131,11 @@ export class CompanyInfo {
      * Get company commerce profile
      *
      * @remarks
-     * Retrieve information about the company, as seen in the commerce platform.
+     * Retrieve information about the company, as seen in the commerce
+     * platform.
      *
-     * This may include information like addresses, tax registration details and social media or website information.
+     * This may include information like addresses, tax registration details and
+     * social media or website information."
      */
     async getCommerceProfile(
         req: operations.GetCommerceProfileRequest,
@@ -220,7 +222,7 @@ export class CompanyInfo {
                     );
                 }
                 break;
-            case [401, 404, 409, 429].includes(httpRes?.status):
+            case [401, 402, 403, 404, 409, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),