+++ lending/src/sdk/financialstatementsprofitandloss.ts
@@ -108,7 +108,7 @@ export class FinancialStatementsProfitAndLoss {
                     );
                 }
                 break;
-            case [401, 404, 409, 429].includes(httpRes?.status):
+            case [401, 402, 403, 404, 409, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -222,7 +222,7 @@ export class FinancialStatementsProfitAndLoss {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),