+++ lending/src/sdk/managedatapulloperations.ts
@@ -106,7 +106,7 @@ export class ManageDataPullOperations {
                     );
                 }
                 break;
-            case [401, 404, 429].includes(httpRes?.status):
+            case [401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -130,15 +130,7 @@ export class ManageDataPullOperations {
      * List pull operations
      *
      * @remarks
-     * The *List pull operations* endpoint returns a list of [pull operations](https://docs.codat.io/lending-api#/schemas/PullOperation) made by your client.
-     *
-     * A [pull operation](https://docs.codat.io/lending-api#/schemas/PullOperation) is a request to retrieve a specific data type from an integration.
-     *
-     * ### Tips and traps
-     *
-     * - The *List pull operations* endpoint does not support querying the `isCompleted` property. You can filter failed pull operations by querying `status!=Complete&&status!=NotSupported` instead.
-     *
-     *
+     * Gets the pull operation history (datasets) for a given company.
      */
     async list(
         req: operations.ListPullOperationsRequest,
@@ -222,7 +214,7 @@ export class ManageDataPullOperations {
                     );
                 }
                 break;
-            case [400, 401, 404, 429].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),