+++ lending/src/sdk/models/operations/createbanktransactions.ts
@@ -55,7 +55,7 @@ export class CreateBankTransactionsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;