+++ lending/src/sdk/models/operations/getbankingtransactioncategory.ts
@@ -40,6 +40,12 @@ export class GetBankingTransactionCategoryResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
+
+    /**
      * HTTP response status code for this operation
      */
     @SpeakeasyMetadata()