+++ lending/src/sdk/models/operations/getcategorizedbankstatement.ts
@@ -46,7 +46,7 @@ export class GetCategorizedBankStatementResponse extends SpeakeasyBase {
     enhancedCashFlowTransactions?: shared.EnhancedCashFlowTransactions;
 
     /**
-     * Your API request was not properly authorized.
+     * Your `query` parameter was not correctly formed
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;