+++ lending/src/sdk/models/operations/getcommercelifetimevaluemetrics.ts
@@ -64,7 +64,7 @@ export class GetCommerceLifetimeValueMetricsResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;