+++ lending/src/sdk/models/operations/getdataintegritysummaries.ts
@@ -40,7 +40,7 @@ export class GetDataIntegritySummariesResponse extends SpeakeasyBase {
     dataIntegritySummaries?: shared.DataIntegritySummaries;
 
     /**
-     * Your API request was not properly authorized.
+     * Your `query` parameter was not correctly formed
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;