+++ lending/src/sdk/salesmetrics.ts
@@ -136,7 +136,7 @@ export class SalesMetrics {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -270,7 +270,7 @@ export class SalesMetrics {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -405,7 +405,7 @@ export class SalesMetrics {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),