+++ lending/src/sdk/salesreports.ts
@@ -131,7 +131,7 @@ export class SalesReports {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),
@@ -267,7 +267,7 @@ export class SalesReports {
                     );
                 }
                 break;
-            case [401, 404].includes(httpRes?.status):
+            case [400, 401, 402, 403, 404, 429, 500, 503].includes(httpRes?.status):
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.errorMessage = utils.objectToClass(
                         JSON.parse(decodedRes),