+++ lending/src/sdk/sdk.ts
@@ -69,9 +69,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "3.4.0";
-    genVersion = "2.169.0";
-    userAgent = "speakeasy-sdk/typescript 3.4.0 2.169.0 3.0.0 @codat/lending";
+    sdkVersion = "3.5.0";
+    genVersion = "2.173.0";
+    userAgent = "speakeasy-sdk/typescript 3.5.0 2.173.0 3.0.0 @codat/lending";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);