+++ lending/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v3.4.0] lending
 ### Releases
 - [NPM v3.4.0] https://www.npmjs.com/package/@codat/lending/v/3.4.0 - lending
+
+## 2023-10-30 01:27:27
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.5.0] lending
+### Releases
+- [NPM v3.5.0] https://www.npmjs.com/package/@codat/lending/v/3.5.0 - lending